Add a 'longest line' calculator for minimap.vim background work
As part of https://github.com/wfxr/minimap.vim/issues/160, it is required to
have an external program that will scan a file for the longest line.
code-minimap is already a dependency for minimap.vim, so tucking the background
work into this binary seems a reasonable way to do that.